requirements.txt
GitPython==3.1.42
torch==2.2.1
transformers==4.38.2
